update eps 30 menu links

+++ app/Http/Controllers/IdeaController.php
@@ -16,18 +16,21 @@ class IdeaController extends Controller
     }
     public function store(){
 
-        request()->validate([
+        $validated = request()->validate([
             'content' => 'required|max:255'
         ]);
 
-        $idea = Idea::create([
-            'content' => request()->get('content'),
-        ]);
+        $validated['user_id'] = auth()->user()->id;
+
+        Idea::create($validated);
 
         return redirect()->route('dashboard')->with('success','Idea created successfully');
     }
 
     public function edit(Idea $idea){
+        if(auth()->id() !== $idea->user_id){
+            abort(404);
+        }
         $editing = true;
 
         return view("ideas.show", compact("idea", "editing"));
@@ -43,6 +46,10 @@ class IdeaController extends Controller
     }
 
     public function destroy(Idea $idea){
+
+        if(auth()->id() !== $idea->user_id){
+            abort(404);
+        }
         $idea->delete();
 
         return redirect()->route('dashboard')->with('success','Idea deleted successfully!');
